tests: add swiss cities
